[258067]  Provide RequestWithMonitoring example https://bugs.eclipse.org/bugs/show_bug.cgi?id=258067

+++ plugins/org.eclipse.net4j.tests/src/org/eclipse/net4j/tests/SignalMonitorTest.java
@@ -176,7 +176,7 @@ public class SignalMonitorTest extends AbstractTransportTest
     }
 
     @Override
-    public void worked(int work)
+    public void worked(double work)
     {
       super.worked(work);
       System.out.println("work: " + getWork());
